server/requirements.txt
Django==1.8.4
django-angular==0.7.15
django-rest-auth==0.5.0
django-taggit==0.17.1
djangorestframework==3.2.3
facebook-sdk==0.4.0
Pillow==2.9.0
psycopg2==2.6.1
pytz==2015.4
six==1.9.0
wheel==0.24.0
